Avoid modifying rebuild buckets state in no_grad context
See https://github.com/pytorch/pytorch/issues/54059 for discussion.

In short, users might want to run evaluation on a single rank
in `torch.no_grad()` mode. When this happens, we need to make
sure that we skip all rebuild bucket logics, as the forward only
runs on one rank and not all peers can sure the bucket configuration
sync communication.

ghstack-source-id: 2c6721c1f8a04002ad24bcf9d5ac2e484de01e14
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54159